fix(wasm): resolve Windows fuel-trap abort and u32/usize limit mismatch
- StoreLimitsBuilder::table_elements takes usize; cast the u32 constant.
- Upgrade wasmtime 26.0.1 -> 29.0.1: 26.x raised OutOfFuel through an
  extern "C" libcall wrapper on Windows, producing a non-unwinding panic
  (STATUS_STACK_BUFFER_OVERRUN) instead of a normal trap. The 29.x
  libcall wrappers are extern "C-unwind", fixing the abort.
- Drop the no-longer-needed wasm_backtrace(false) workaround.

+++ crates/daemon/src/wasm_sandbox.rs
@@ -22,7 +22,6 @@ struct SandboxState {
 }
 
 /// Builds the shared engine used by every Steward WASM entry point. Fuel consumption must be
-/// enabled on the engine before stores can receive an execution budget.
 pub fn build_engine() -> Result<Engine> {
     let mut config = Config::new();
     config.consume_fuel(true);
@@ -42,7 +41,7 @@ pub fn execute(engine: &Engine, bytes: &[u8]) -> Result<i32> {
     let module = Module::from_binary(engine, bytes).context("compiling WASM")?;
     let limits = StoreLimitsBuilder::new()
         .memory_size(MAX_WASM_MEMORY_BYTES)
-        .table_elements(MAX_WASM_TABLE_ELEMENTS)
+        .table_elements(MAX_WASM_TABLE_ELEMENTS as usize)
         .instances(MAX_WASM_INSTANCES)
         .memories(MAX_WASM_MEMORIES)
         .tables(MAX_WASM_TABLES)
